(CIL) fix Metadata decoding failure
decoding of Param was declared to use the Blob heap for names,
but needs to use the String heap. On some binaries this created
errors because of different index widths (16 vs 32 bit) depending
on the heap sizes.

Issue #471

+++ src/AstNodes/BinaryAnalysis/SgAsmCilMethodDef.h
@@ -12,7 +12,7 @@ class SgAsmCilMethodDef: public SgAsmCilMetadata {
      *
      *  A 2-byte bitmask of type MethodImplAttributes. */
     [[using Rosebud: rosetta]]
-    uint32_t ImplFlags = 0;
+    uint16_t ImplFlags = 0;
 
     /** Property: Flags.
      *
